Move CalcShaderUniforms to FramebufferCommon
